Removed min 2 file limitation
The app now has the functionality of reducing size. Scenario: there is a size constraint in sending/submitting the image, but you only find out later. Adding the picture back, pressing generate and reducing the size is now a viable option; you don't need to add the seperate images again.