Added DPI correction
The image now compensates for non-96 DPI images, which resulted in the images being scaled.
E.g. a 96 DPI image had 1 pixel input 1 pixel output ratio, but a 182 DPI would have had half the size at output, resulting in images scaling "randomly" over the place.
As a sidenote: the method PreviewForm.ResizeImage(...) was modified, because my first approach was a much more complicated image rescaling, rerendering etc one, but turns out it  can be solved in an easier way. Actually the two methods are quite similar, but this seemed cleaner to me. The same reasoning applies to the ArrangerForm.ChangeDPI(...) method