Fixed file locking issue
The files are released from memory if they are removed from the working space.
The files are loaded in a different way by default, which leaves out certain markers and metadata, but does not seem to effect normal usage. DPI is still copied.
A "Lock-Load Files" option has been added, should this cause a problem.

Also: the image number check has been reintroduced to prevent freeze on 0 image generation